added unit test of diff cent to state
