+++ src-tauri/src/analytics/algorithms/diffcen.rs
@@ -63,31 +63,3 @@ pub fn diffusion_centrality(
 
     node_to_diffcen
 }
-
-// add unit tests
-/* #[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn test_diffusion_centrality() {
-        let mut G1 = Graph::new();
-
-        // Create a few nodes and edges and add to graph
-        let u: String = "u".to_string();
-        let v: String = "v".to_string();
-        let w: String = "w".to_string();
-
-        let _u_idx = G1.add_node(u.clone());
-        let _v_idx = G1.add_node(v.clone());
-        let _w_idx = G1.add_node(w.clone());
-
-        G1.add_edge(u.clone(), v.clone(), 1 as f64);
-        G1.add_edge(u.clone(), w.clone(), 7 as f64);
-        G1.add_edge(v.clone(), w.clone(), 35 as f64);
-
-        let diff_cents = diffusion_centrality(&G1, 1);
-        println!("{:?}", diff_cents);
-        assert!(diff_cents.is_some());
-    }
-} */
